Session Management Adjustments
Following changes are to protect ensure that page sessions are released as soon as they are no longer needed. Sessions that remain locked as writable can cause php-pfm threads to hang while it waits for the session of the previous page call to be released.

Only create a session when needed, such has when a user is logged in, preforming an occurrence search, or triggering a download
Run session_write_close() function when session is no longer needed